moved logic to cardform
